feat(): add basic applet
Create basic utilities for creating tests, and examining and display
their results.

Currently you need to create an instance of DBTests class around a
pandas dataframe, and then you can add preficates which preform tests on
a given row to check the validity of a specified column.

Calling run() on the instance will produce a summary of the coverage
across columns, the success of each individual test across its column
and a sample of the places where it fails.