fix(dftest): no --files returns None, not empty list
previously test files were specified as a positional argument, meaning not specifying files would result in the corresponding attribute being an empty list. Since it was converted to the option --files, it returns None if not specified.

This also changes the number of arguments from '*' to '+', i.e must have at least 1 argument, which wasn't so previously as a way to make it an optional positional argument